int array change

+++ light_map.ino
@@ -1,5 +1,5 @@
-float light_map(int raw)  {
-  float light = map(raw, 0, 800, 0, 100);
+int light_map(int raw)  {
+  int light = map(raw, 0, 800, 0, 100);
   
   return light;  
 }
