+++ running_avg.ino
@@ -1,7 +1,7 @@
-float running_avg(float* array, int count)  {
+int running_avg(int* array, int count)  {
   int avg_start = 0;
   float sum = 0;
-  float num_el = float(count);
+  int num_el = float(count);
  
   if(count >= 20)  {  
     avg_start = count - 20;
@@ -13,5 +13,5 @@ float running_avg(float* array, int count)  {
   
   float average = sum/num_el;
   
-  return average;
+  return int(average);
 }
